+++ src/shared/components/data-table/components/main/components/table-body/TableBody.tsx
@@ -14,33 +14,15 @@
  * limitations under the License.
  */
 
-import React, { useContext } from 'react';
+import React from 'react';
 
-import { TableContext } from 'src/shared/components/data-table/DataTable';
 import TableRow from '../table-row/TableRow';
 
-import { defaultDataTableState } from 'src/shared/components/data-table/dataTableReducer';
-
-import {
-  DataTableColumns,
-  DataTableState,
-  SortingDirection
-} from 'src/shared/components/data-table/dataTableTypes';
+import useDataTable from 'src/shared/components/data-table/hooks/useDataTable';
 
 const TableBody = () => {
-  const {
-    data,
-    currentPageNumber,
-    rowsPerPage,
-    columns,
-    sortedColumn,
-    uniqueColumnId,
-    hiddenRowIds
-  } = useContext(TableContext) || {
-    currentPageNumber: defaultDataTableState.currentPageNumber,
-    rowsPerPage: defaultDataTableState.rowsPerPage,
-    hiddenRowIds: {}
-  };
+  const { getSortedCurrentPageRows, data, columns, uniqueColumnId } =
+    useDataTable();
 
   if (!(data && columns && uniqueColumnId)) {
     return null;
@@ -50,99 +32,18 @@ const TableBody = () => {
   );
 
   // Filter the rows that needs to be displayed in the current page
-  const rowsThisPage = getCurrentPageRows({
-    hiddenRowIds,
-    data,
-    currentPageNumber,
-    uniqueColumnId,
-    rowsPerPage,
-    columns
-  });
-
-  if (sortedColumn) {
-    const sortedColumnIndex = columns.findIndex(
-      (column) => column.columnId === sortedColumn.columnId
-    );
-    rowsThisPage.sort((currentRow, nextRow) => {
-      const currentValue = currentRow[sortedColumnIndex]?.toString() || '';
-      const nextValue = nextRow[sortedColumnIndex]?.toString() || '';
-
-      const currentValueAsNumber = Number(currentValue);
-      const nextValueAsNumber = Number(nextValue);
-      if (!isNaN(currentValueAsNumber) && !isNaN(nextValueAsNumber)) {
-        if (currentValueAsNumber < nextValueAsNumber) {
-          return sortedColumn.sortedDirection === SortingDirection.ASC ? -1 : 1;
-        } else if (currentValueAsNumber > nextValueAsNumber) {
-          return sortedColumn.sortedDirection === SortingDirection.ASC ? 1 : -1;
-        }
-      } else {
-        if (currentValue < nextValue) {
-          return sortedColumn.sortedDirection === SortingDirection.ASC ? -1 : 1;
-        } else if (currentValue > nextValue) {
-          return sortedColumn.sortedDirection === SortingDirection.ASC ? 1 : -1;
-        }
-      }
-
-      return 0;
-    });
-  }
+  const rowsThisPage = getSortedCurrentPageRows();
 
   return (
     <tbody>
       {rowsThisPage.map((rowData, index) => {
         const rowId = String(rowData[uniqueColumnIndex]);
-
         return (
           <TableRow key={index} rowData={rowData} rowId={rowId as string} />
         );
       })}
     </tbody>
   );
-};
-
-type GetCurrentPageRowsParams = Pick<
-  DataTableState,
-  'hiddenRowIds' | 'data' | 'currentPageNumber' | 'rowsPerPage'
-> & {
-  uniqueColumnId: string;
-  columns: DataTableColumns;
-};
-
-export const getCurrentPageRows = (params: GetCurrentPageRowsParams) => {
-  const {
-    hiddenRowIds,
-    data,
-    currentPageNumber,
-    uniqueColumnId,
-    rowsPerPage,
-    columns
-  } = params;
-
-  const uniqueColumnIndex = columns.findIndex(
-    (column) => column.columnId === uniqueColumnId
-  );
-
-  const totalRows = data.length;
-  const rowIndexLowerBound = (currentPageNumber - 1) * rowsPerPage;
-  const rowIndexUpperBound = rowIndexLowerBound + rowsPerPage;
-
-  const visibleRows = hiddenRowIds
-    ? data.filter((rowData) => {
-        const rowId = String(rowData[uniqueColumnIndex]);
-
-        return hiddenRowIds[rowId] !== true;
-      })
-    : data;
-
-  // Filter the rows that needs to be displayed in the current page
-  return visibleRows.filter((_, rowIndex) => {
-    if (totalRows > rowsPerPage) {
-      if (rowIndex < rowIndexLowerBound || rowIndexUpperBound - 1 < rowIndex) {
-        return false;
-      }
-    }
-    return true;
-  });
 };
 
 export default TableBody;
